The Employee Logic

+++ src/main/java/com/LogisticsCompany/dto/EmployeeDTO.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class EmployeeDTO {
     private String name;
-    private double salary;
+    private BigDecimal salary;
     private Long officeID;
     private EmployeeType employeeType;
 }
